Move CLI parsing utils to separate module
This change was introduced in the 1.9.54 SDK.
